requirements.txt
requests==2.29
django==4.2
pytest==7.3.1
pytest-django==4.5.2
Pillow==9.5.0
PyJWT==2.6.0
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
